refacttoring the code
